feat: add logs.h macros as logger

+++ utils.c
@@ -1,7 +1,8 @@
 #include <stdio.h>
+#include <stdint.h>
 #include <stdlib.h>
 
 void fill_buffer_with_random_bytes(char *buffer) {
-	for (int i = 0; i < sizeof(buffer); i++)
+	for (uint16_t i = 0; i < sizeof(buffer); i++)
         	buffer[i] = rand() % 256;
 }
